requirements.txt
fastapi
uvicorn
supabase
google-generativeai
python-dotenv
tavily-python
